[language] Cleanup: remove execute_function_in_module from the runtime
execute_function_in_module is used by the test generation only and it's a wrapper that
should live outside of the VM.
